+++ src/types/api-response.ts
@@ -11,13 +11,12 @@ export type ObjectKind =
 
 export type ApiResponse = {
   object: ObjectKind;
+  // TODO can we get away with unknown here?
+  // or not having it at all?
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   [property: string]: any;
 };
 
-export interface CardApiResponse extends ApiResponse {
-  object: "card";
-}
 export interface ListApiResponse extends ApiResponse {
   object: "list";
   data: ApiResponse[];
@@ -25,7 +24,4 @@ export interface ListApiResponse extends ApiResponse {
 export interface CatalogApiResponse extends ApiResponse {
   object: "catalog";
   data: string[];
-}
-export interface SetApiResponse extends ApiResponse {
-  object: "set";
 }